add report controler

+++ Benaa.Core/Interfaces/IRepositories/IWalletRepository.cs
@@ -5,7 +5,7 @@ namespace Benaa.Core.Interfaces.IRepositories
     public interface IWalletRepository : IBaseRepository<Wallet>
     {
         Task<int> GetAmountCode(string code);
-        Task<decimal> AddAmountCode(string ui, int amount);
+        Task<decimal> AddAmountCode(string ui, decimal amount);
         Task<decimal> check(string u);
 
         Task<string> getTecherid(Guid itemID);
